+++ src/orders/models/__init__.py
@@ -1,9 +1,9 @@
 from orders.models.order import Order
 from orders.models.promocode import PromoCode
-from orders.models.refund import Refund
+from orders.models.refund_log_entry import RefundLogEntry
 
 __all__ = [
     "Order",
     "PromoCode",
-    "Refund",
+    "RefundLogEntry",
 ]
