--- src/orders/models/refund.py
@@ -1,18 +1,0 @@
-from django.db import models
-
-from app.models import TimestampedModel
-
-
-class Refund(TimestampedModel):
-    order = models.ForeignKey("orders.Order", on_delete=models.CASCADE, related_name="refunds")
-    author = models.ForeignKey("users.User", on_delete=models.RESTRICT, related_name="refunds")
-    author_ip = models.GenericIPAddressField()
-    bank_confirmation_received = models.BooleanField(default=False, verbose_name="Bank refund confirmation")
-
-    class Meta:
-        verbose_name = "Refund"
-        verbose_name_plural = "Refunds"
-        ordering = ["-created"]
-
-    def __str__(self) -> str:
-        return f"Refund for order #{self.order.pk}, by {self.author}"
