+++ src/orders/services/__init__.py
@@ -1,10 +1,10 @@
+from orders.services.order_bank_refunder import OrderBankRefunder
 from orders.services.order_creator import OrderCreator
 from orders.services.order_diploma_generator import OrderDiplomaGenerator
 from orders.services.order_email_changer import OrderEmailChanger
 from orders.services.order_paid_setter import OrderPaidSetter
-from orders.services.order_refunder import OrderRefunder
+from orders.services.order_refunder import OrderUnpaidSetter
 from orders.services.order_shipper import OrderShipper
-from orders.services.order_unpaid_setter import OrderUnpaidSetter
 from orders.services.order_unshipper import OrderUnshipper
 
 __all__ = [
@@ -15,5 +15,5 @@ __all__ = [
     "OrderUnpaidSetter",
     "OrderShipper",
     "OrderUnshipper",
-    "OrderRefunder",
+    "OrderBankRefunder",
 ]
